Moved tvb reader to io package.
